pyproject.toml
[tool.poetry]
name = "goit-python-web-fastapi-lect-01"
version = "0.2.0"
description = ""
authors = ["lexxai <lexxai@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
pydentic = {extras = ["email"], version = "^0.0.1.dev3"}
pydantic = {extras = ["email"], version = "^2.4.2"}
psycopg2 = "^2.9.9"
alembic = "^1.12.1"
jinja2 = "^3.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
